Moving encoding related code to encoder
